requirements.txt
Flask
groq
PyPDF2
python-docx